refactor(editor): extract hierarchy navigation controller
Move per-Cell camera restoration, caller-path selection, locator and diagnostic focus, hierarchy enter/up/top commands, and rectangle-to-Cell entry out of App into the Hierarchy feature.

Validation: pnpm typecheck; pnpm ci:static; 22 focused hierarchy/session/rectangle/toolbar/App unit tests; focused manual-editor Playwright contracts for Project Search, ERC focus, and visual diagnostic navigation; focused hierarchy Playwright contracts for hierarchy visibility, caller navigation, and shared Cell placement; pnpm gate:plan for owned paths; git diff --check. Parallel browser startup initially collided on port 4173; all affected cases passed when rerun serially.

Test-Impact: no-test-change — behavior is unchanged and covered by existing hierarchy/navigation units and focused browser navigation contracts.
